backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
pydantic-settings==2.6.1
alembic==1.13.3
python-multipart==0.0.12
httpx==0.27.2
beautifulsoup4==4.12.3
pypdf==5.1.0
PyJWT==2.9.0
androguard==4.1.2
